updated style
