+++ src/plot_page/__init__.py
@@ -1,13 +1,13 @@
 from dash import Input, Output, State
 
-from plot_page.view.components.app import app
-from plot_page.control.data_operations import prepare_upload_data
+from plot_page.control.data_operation.management_data import prepare_upload_data
 from plot_page.view.components import page_layout  # noqa: F811
-from plot_page.view.pages import plot_2d
+from plot_page.view.components.app import app
 from plot_page.view.pages.data_analyse import data_analyse_layout
 
 
 from plot_page.view.pages.upload import upload_layout
+from plot_page.view.pages.visualization.plot2d import plot2d_layout
 
 if __name__ == "__main__":
     app.layout = page_layout()
@@ -30,7 +30,7 @@ def display_page(pathname: str) -> list:
         tuple[html.Div, dict]: List with html components for page-content and data.
     """
     if pathname == "/plot_2d":
-        return plot_2d.layout()
+        return plot2d_layout()
     if pathname == "/data_analyse":
         return data_analyse_layout()
     else:
