+++ src/plot_page/control/data_operation/__init__.py
@@ -1,0 +1,1 @@
+"""This module contains function for operations on the data."""
