+++ src/plot_page/control/visualisation/__init__.py
@@ -1,0 +1,1 @@
+"""Module with function that are connected to visualisation."""
