+++ src/plot_page/data/__init__.py
@@ -1,0 +1,1 @@
+"""This module is used for storing data."""
