--- src/plot_page/data/global_variables.py
@@ -1,5 +1,0 @@
-import os
-
-
-DATAFRAME_STORE = os.path.join(".", "Data")
-os.makedirs(DATAFRAME_STORE, exist_ok=True)
