+++ src/plot_page/view/components/app.py
@@ -1,3 +1,5 @@
+"""App components that need to be used by all pages."""
+
 import dash_bootstrap_components as dbc
 from dash import Dash
