+++ src/plot_page/view/pages/visualization/__init__.py
@@ -1,0 +1,1 @@
+"""Contains multiple pages that are use for the visualization of the data."""
